+++ src/dockerValidator.ts
@@ -6,7 +6,7 @@ import {
     TextDocument, Range, Position, Diagnostic, DiagnosticSeverity
 } from 'vscode-languageserver-types';
 import { Dockerfile, Flag, Instruction, Env, Label, Onbuild, ModifiableInstruction, DockerfileParser, Directive } from 'dockerfile-ast';
-import { ValidationSeverity, ValidatorSettings } from './main';
+import { ValidationCode, ValidationSeverity, ValidatorSettings } from './main';
 
 export const KEYWORDS = [
     "ADD",
@@ -28,52 +28,6 @@ export const KEYWORDS = [
     "VOLUME",
     "WORKDIR"
 ];
-
-export enum ValidationCode {
-    CASING_INSTRUCTION,
-    CASING_DIRECTIVE,
-    ARGUMENT_MISSING,
-    ARGUMENT_EXTRA,
-    ARGUMENT_REQUIRES_ONE,
-    ARGUMENT_REQUIRES_AT_LEAST_ONE,
-    ARGUMENT_REQUIRES_TWO,
-    ARGUMENT_REQUIRES_AT_LEAST_TWO,
-    ARGUMENT_REQUIRES_ONE_OR_THREE,
-    ARGUMENT_UNNECESSARY,
-    DUPLICATE_BUILD_STAGE_NAME,
-    EMPTY_CONTINUATION_LINE,
-    INVALID_BUILD_STAGE_NAME,
-    FLAG_AT_LEAST_ONE,
-    FLAG_DUPLICATE,
-    FLAG_INVALID_DURATION,
-    FLAG_LESS_THAN_1MS,
-    FLAG_MISSING_DURATION,
-    FLAG_MISSING_VALUE,
-    FLAG_UNKNOWN_UNIT,
-    NO_SOURCE_IMAGE,
-    INVALID_ESCAPE_DIRECTIVE,
-    INVALID_AS,
-    INVALID_PORT,
-    INVALID_PROTO,
-    INVALID_SIGNAL,
-    INVALID_SYNTAX,
-    ONBUILD_CHAINING_DISALLOWED,
-    ONBUILD_TRIGGER_DISALLOWED,
-    SHELL_JSON_FORM,
-    SHELL_REQUIRES_ONE,
-    SYNTAX_MISSING_EQUALS,
-    SYNTAX_MISSING_NAMES,
-    SYNTAX_MISSING_SINGLE_QUOTE,
-    SYNTAX_MISSING_DOUBLE_QUOTE,
-    MULTIPLE_INSTRUCTIONS,
-    UNKNOWN_INSTRUCTION,
-    UNKNOWN_ADD_FLAG,
-    UNKNOWN_COPY_FLAG,
-    UNKNOWN_HEALTHCHECK_FLAG,
-    UNKNOWN_TYPE,
-    DEPRECATED_MAINTAINER,
-    HEALTHCHECK_CMD_ARGUMENT_MISSING
-}
 
 export class Validator {